Edits Based on Code Failure and CDS Beta Update
In script 1, edited request syntax based on updates to CDS Beta query and ecmwfr R package. Also, edited to remove Kenya-specific language.

In script 2, made minor changes to be in line with less Kenya-specific language

In script 3, modified based on errors in previous code, and changes in the storage of time data in ERA5 SpatRaster, and to reduce Kenya-specific language